backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
motor>=3.7.0
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
playwright>=1.41.0